added most of the necessary backend files for users login.

+++ src/main/java/com/schedule_project/enrolled/Enrolled.java
@@ -57,7 +57,7 @@ public class Enrolled implements Serializable {
     }
 
     @Transient
-    public Integer getStudent() {
+    public Long getStudent() {
         return student.getStudentId();
     }
